rename WRS -> WeightedRandomSampling
