+++ src/Common/tests/gtest_weighted_random_sampling.cpp
@@ -1,4 +1,4 @@
-#include <Common/WRS_algorithm.h>
+#include <Common/WeightedRandomSampling.h>
 
 #include <gtest/gtest.h>
